Add KV lookup for RFCs, modify 302s for I-Ds
When a request hits this worker, we check the cache, via URL, to see
if there is an already-transformed document entry. With this change,
if there is a cache miss, we'll check if there is a canonical entry
stored in the KV store bound to `IETF`.

Furthermore, if a request points to an unversioned I-D,
the origin returns a redirect to the latest version hosted on tools.ietf.org.
This sidestepped the worker and its optimizations. This change rewrites
the Location header and content body to point to the worker.

With this change, we have several layers of caching and retrieval.
To help communicate the worker actions to external observers, we
add a new "edge-tools-ietf" headr that states where the response
payload came from.